Minor cleanup/refactor
