+++ harmony-fe/src/components/post-modal/topic-field.tsx
@@ -5,7 +5,8 @@ import { Topic } from '../../models/topic'
 import useHttpRequest, { HttpMethod } from '../../hooks/httpRequest'
 import { PostField } from '../../models/post'
 import AutocompleteLi from '../autocomplete-li/autocomplete-li'
-import { TSearchOption } from '../search-bar/search-bar'
+import { getSearchableLabel } from '../../utils/additionalContext'
+import { Searchable } from '../../models/searchable'
 
 interface ITopicFieldProps {
   error?: boolean
@@ -48,7 +49,7 @@ export const TopicField = ({ error, onChange }: ITopicFieldProps) => {
       options={topicOptions}
       onInputChange={(event, value) => onTopicInputChange(value)}
       onChange={(event, value) => onTopicChange(value)}
-      getOptionLabel={topic => topic?.name || ''}
+      getOptionLabel={option => getSearchableLabel(option as Searchable)}
       isOptionEqualToValue={(option, value) => option?.name === value?.name}
       fullWidth
       autoHighlight
@@ -65,7 +66,7 @@ export const TopicField = ({ error, onChange }: ITopicFieldProps) => {
         />
       )}
       renderOption={(props, option) => (
-        <AutocompleteLi option={option as TSearchOption} {...props} />
+        <AutocompleteLi option={option as Searchable} {...props} />
       )}
     />
   )
