+++ harmony-fe/src/models/searchable.ts
@@ -1,0 +1,4 @@
+import { Topic } from './topic'
+import { User } from './user'
+
+export type Searchable = Partial<User> | Topic
